Fixed issue where the paicoin-cli -testnet still loaded the mainnet.
I wrapped paicoin with bitcoin's test framework and apparently having
a directory named testnet, where testnet had been ran privately is not
sufficient using this wrapper. That is the wrapper checks the associated
.conf file for whether or not testnet=1 (or regtest=1 for regtest) and
correspondingly decides whether or not it is a testnet node. I put that
explicitly in there and it now seems to work.